demo/requirements.txt
torch==2.5.1
torchvision
torchaudio

accelerate==1.1.1
gradio
gradio_client
spaces

Pillow==11.1.0
opencv-python-headless==4.11.0.86
datasets==3.6.0

transformers==4.51.3
qwen-vl-utils==0.0.8

pre-commit==4.2.0
matplotlib
flash-attn==2.7.3